rename: features 폴더에 event 도메인끼리 묶기
