--- src/features/event-manage/event-create/model/event.ts
@@ -1,7 +1,0 @@
-import { BaseEvent } from '../../../../shared/types/baseEventType';
-
-export interface CreateEventRequest extends BaseEvent {
-  hostChannelId: number;
-}
-
-
