--- src/features/event-manage/event-list/model/eventList.ts
@@ -1,7 +1,0 @@
-import { BaseEvent } from '../../../../shared/types/baseEventType';
-
-export interface EventList extends BaseEvent {
-  id: number;
-  hostChannelName: string;
-  remainDays: string;
-}
