+++ src/features/event/hooks/useEventHook.ts
@@ -1,5 +1,5 @@
 import { AxiosError } from 'axios';
-import { ApiResponse } from '../../../../shared/types/api/apiResponse';
+import { ApiResponse } from '../../../shared/types/api/apiResponse';
 import { createEvent } from '../api/event';
 import { CreateEventRequest } from '../model/event';
 import { useMutation } from '@tanstack/react-query';
