+++ src/features/event/hooks/useHostHook.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { createHost, deleteHost } from '../api/host';
 import { HostCreationRequest } from '../model/hostCreation';
-import { ApiResponse } from '../../../../shared/types/api/apiResponse';
+import { ApiResponse } from '../../../shared/types/api/apiResponse';
 
 export const useHostCreation = () => {
   return useMutation<ApiResponse<null>, Error, HostCreationRequest>({
