+++ src/features/event/ui/EventFunnel.tsx
@@ -1,14 +1,14 @@
-import HostSelectionPage from '../../../../pages/event-manage/ui/HostSelectionPage';
-import HostCreationPage from '../../../../pages/event-manage/ui/HostCreationPage';
-import EventTitlePage from '../../../../pages/event-manage/ui/EventTitlePage';
-import EventPeriodPage from '../../../../pages/event-manage/ui/EventPeriodPage';
-import EventInfoPage from '../../../../pages/event-manage/ui/EventInfoPage';
-import EventTypePage from '../../../../pages/event-manage/ui/EventTypePage';
-import EventTagPage from '../../../../pages/event-manage/ui/EventTagPage';
-import EventOrganizerInfoPage from '../../../../pages/event-manage/ui/EventOrganizerInfoPage';
-import EventRegisterLayout from '../../../../shared/ui/backgrounds/EventRegisterLayout';
+import HostSelectionPage from '../../../pages/event-manage/ui/HostSelectionPage';
+import HostCreationPage from '../../../pages/event-manage/ui/HostCreationPage';
+import EventTitlePage from '../../../pages/event-manage/ui/EventTitlePage';
+import EventPeriodPage from '../../../pages/event-manage/ui/EventPeriodPage';
+import EventInfoPage from '../../../pages/event-manage/ui/EventInfoPage';
+import EventTypePage from '../../../pages/event-manage/ui/EventTypePage';
+import EventTagPage from '../../../pages/event-manage/ui/EventTagPage';
+import EventOrganizerInfoPage from '../../../pages/event-manage/ui/EventOrganizerInfoPage';
+import EventRegisterLayout from '../../../shared/ui/backgrounds/EventRegisterLayout';
 import { useNavigate } from 'react-router-dom';
-import { EventFunnelInterface, StepNames } from '../../../../shared/types/funnelType';
+import { EventFunnelInterface, StepNames } from '../../../shared/types/funnelType';
 import { useFunnelState } from '../model/FunnelContext';
 import { useEventCreation } from '../hooks/useEventHook';
 import { useHostCreation } from '../hooks/useHostHook';
