+++ src/features/event/ui/ShareEventModal.tsx
@@ -1,7 +1,7 @@
-import profile from '../../../../../public/assets/banners/1.png';
-import link from '../../../../../public/assets/event-manage/details/Link.svg';
-import kakao from '../../../../../public/assets/event-manage/details/KaKao.svg';
-import { shareToKakao } from '../../../../shared/lib/kakaoShare';
+import profile from '../../../../public/assets/banners/1.png';
+import link from '../../../../public/assets/event-manage/details/Link.svg';
+import kakao from '../../../../public/assets/event-manage/details/KaKao.svg';
+import { shareToKakao } from '../../../shared/lib/kakaoShare';
 
 interface ShareEventModalProps {
   closeModal: () => void;
