+++ src/pages/event-manage/ui/EventTitlePage.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import UnderlineTextField from '../../../../design-system/ui/textFields/UnderlineTextField';
-import { useFunnelState } from '../../../features/event-manage/event-create/model/FunnelContext';
+import { useFunnelState } from '../../../features/event/model/FunnelContext';
 import { EventTitleFormData, eventTitleZodValidation } from '../../../shared/lib/formValidation';
 
 interface EventTitlePageProps {
